requirements.txt
# Pictures - Phase 1 Requirements (Flexible Versions)
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
aiohttp>=3.8.0
openai>=1.0.0
runwayml>=3.0.0
replicate>=0.20.0
asyncpg>=0.28.0
pillow>=10.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0

# Sprint 3: Production Infrastructure
redis>=4.5.0
redis[hiredis]>=4.5.0
asyncpg>=0.29.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=23.0.0
flake8>=6.0.0